+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
+## Unreleased
+
+- Add the beta `@depot/sandbox` package under `packages/sandbox`.
+
 See [GitHub Releases](https://github.com/depot/sdk-node/releases) for the latest changes.